requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==12.4.2
psycopg2-binary
feedparser
PyLyrics
tldextract
Pillow>=4.2.0
alphabet_detector
bleach
markdown2
spamwatch
regex
googletrans
